rename component with uppercace
